backend/pyproject.toml
[project]
name = "innovatepam-portal-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn>=0.30.0",
  "sqlalchemy>=2.0.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "email-validator>=2.2.0",
  "pydantic>=2.8.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-mock>=3.14.0",
  "pytest-cov>=5.0.0",
  "httpx>=0.27.0",
  "mypy>=1.10.0",
  "pylint>=3.2.0",
  "types-python-jose>=3.3.4",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
